Add batch cycle support to RISC-V testbenches and core
The loop calling riscv_step() was moved partially out of the testbench to ensure that cosim could run without using hundreds of GB of space and taking hours. This caused a couple of small changes in the current testbenches so they could use this new format. The hardcoded testbench was added for cosim.